refactor: 댓글 수정 영역 컴포넌트로 분리
